Dark mode (#56)
close #53 : new dark mode option can be set !

In the menu bar, there is now the possibility of enabling a dark mode. The change happens dynamically.

New colors, and a dark background for sensitive eyes and for those that prefer the darker look.
Light mode stays the default one, and one just has to uncheck the dark mode setting to revert to it.